# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
  /** AI / xAI service health (circuit breaker state) */
  ai?: {
    /** healthy | degraded | open */
    status: "healthy" | "degraded" | "open";
    consecutiveTransients?: number;
    rateLimitCount?: number;
    otherTransientCount?: number;
    cooldownRemainingMs?: number;
  };
}

export interface Account {
  userId: string;
  /** @nullable */
  email?: string | null;
  isPro: boolean;
  isAdmin: boolean;
  imagesGeneratedThisMonth: number;
  monthlyImageLimit: number;
  topicCredits: number;
  freeBreakdownsUsedThisMonth: number;
  freeBreakdownsPerMonth: number;
  /** @nullable */
  subscriptionStatus?: string | null;
  /** @nullable */
  subscriptionCurrentPeriodEnd?: string | null;
  /** Tiny AI/xAI health summary from circuit breaker (Pro users) */
  aiStatus?: "healthy" | "degraded" | "open";
}

export interface ImageInput {
  /** @minLength 1 */
  prompt: string;
}

export interface GeneratedImage {
  url: string;
  imagesGeneratedThisMonth: number;
}

export interface StripeRedirect {
  url: string;
}

export interface BreakdownInput {
  /** @minLength 1 */
  topic: string;
}

export interface GapsInput {
  /** @minLength 1 */
  topic: string;
  breakdownTitles: string[];
}

export interface FindMoreCompaniesInput {
  /** @minLength 1 */
  topic: string;
  /** @minLength 1 */
  gapTitle: string;
  gapWhyExists: string;
  gapInnovationPotential: string;
  existingTickers: string[];
}

export type FindMoreCompaniesResultCompaniesItem = {
  name: string;
  ticker: string;
  exchange: string;
  relevance: string;
};

export interface FindMoreCompaniesResult {
  companies: FindMoreCompaniesResultCompaniesItem[];
}

export interface StockAnalysisInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  ticker: string;
  /** @minLength 1 */
  exchange: string;
  relevance: string;
  /** @minLength 1 */
  topic: string;
  /** @minLength 1 */
  gapTitle: string;
  gapWhyExists: string;
  gapInnovationPotential: string;
}

export interface StockAnalysis {
  analysis: string;
}

export type StripeCheckoutRequestInterval =
  (typeof StripeCheckoutRequestInterval)[keyof typeof StripeCheckoutRequestInterval];

export const StripeCheckoutRequestInterval = {
  month: "month",
  year: "year",
} as const;

export interface StripeCheckoutRequest {
  interval?: StripeCheckoutRequestInterval;
}

export type CreditCheckoutRequestPack =
  (typeof CreditCheckoutRequestPack)[keyof typeof CreditCheckoutRequestPack];

export const CreditCheckoutRequestPack = {
  NUMBER_1: "1",
  NUMBER_5: "5",
  NUMBER_10: "10",
  NUMBER_20: "20",
} as const;

export interface CreditCheckoutRequest {
  pack: CreditCheckoutRequestPack;
}

export type GenerateProBreakdown200 = { [key: string]: unknown };

export type RegenerateProGaps200GapsItem = { [key: string]: unknown };

export type RegenerateProGaps200 = {
  gaps: RegenerateProGaps200GapsItem[];
  [key: string]: unknown;
};