# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: account
    description: Authenticated account info
  - name: images
    description: Server-side image generation (Pro tier)
  - name: stripe
    description: Stripe checkout & customer portal
  - name: breakdown
    description: Server-side first-principles breakdown (Pro tier)
  - name: stocks
    description: AI-powered stock analysis (Pro tier)
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /me:
    get:
      operationId: getMe
      tags: [account]
      summary: Current user account & plan
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "401":
          description: Unauthorized
  /images:
    post:
      operationId: generateProImage
      tags: [images]
      summary: Generate an image with Grok Imagine (Pro tier required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageInput"
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedImage"
        "401":
          description: Unauthorized
        "402":
          description: Pro tier required
        "429":
          description: Monthly image limit reached
  /breakdown:
    post:
      operationId: generateProBreakdown
      tags: [breakdown]
      summary: Generate a first-principles breakdown for a topic (Pro tier required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BreakdownInput"
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Unauthorized
        "402":
          description: Pro tier required
        "502":
          description: Upstream model error
  /breakdown/gaps:
    post:
      operationId: regenerateProGaps
      tags: [breakdown]
      summary: Regenerate the innovation gaps for a topic (Pro tier required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GapsInput"
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                required: [gaps]
                properties:
                  gaps:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
        "401":
          description: Unauthorized
        "402":
          description: Pro tier required
        "502":
          description: Upstream model error
  /stocks/find-companies:
    post:
      operationId: findMoreCompanies
      tags: [stocks]
      summary: Ask Grok to find additional publicly traded companies working on an innovation gap (Pro tier required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FindMoreCompaniesInput"
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindMoreCompaniesResult"
        "401":
          description: Unauthorized
        "402":
          description: Pro tier required
        "502":
          description: Upstream model error
  /stocks/analyze:
    post:
      operationId: analyzeProStock
      tags: [stocks]
      summary: Generate an AI analysis of a public company in the context of an innovation gap (Pro tier required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockAnalysisInput"
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockAnalysis"
        "401":
          description: Unauthorized
        "402":
          description: Pro tier required
        "502":
          description: Upstream model error
  /stripe/checkout:
    post:
      operationId: createStripeCheckoutSession
      tags: [stripe]
      summary: Create a Stripe Checkout session for the Pro subscription
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StripeCheckoutRequest"
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StripeRedirect"
        "401":
          description: Unauthorized
        "502":
          description: Stripe error
  /stripe/credits/checkout:
    post:
      operationId: createCreditCheckoutSession
      tags: [stripe]
      summary: Create a Stripe Checkout session for a one-time topic-credit pack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditCheckoutRequest"
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StripeRedirect"
        "400":
          description: Invalid pack
        "401":
          description: Unauthorized
        "502":
          description: Stripe error
  /stripe/portal:
    post:
      operationId: createStripePortalSession
      tags: [stripe]
      summary: Create a Stripe Billing Portal session for the current user
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StripeRedirect"
        "401":
          description: Unauthorized
        "404":
          description: No Stripe customer for this account
        "502":
          description: Stripe error
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
          description: Basic liveness status
        ai:
          type: object
          description: AI / xAI service health (circuit breaker state for smarter client behavior)
          properties:
            status:
              type: string
              enum: [healthy, degraded, open]
              description: healthy = normal, degraded = elevated transients, open = circuit breaker active (fast-fail in effect)
            consecutiveTransients:
              type: integer
              description: Number of consecutive transient failures (RateLimit/Timeout/etc.)
            cooldownRemainingMs:
              type: integer
              description: Milliseconds until circuit closes (0 when healthy or degraded)
          required:
            - status
      required:
        - status
    Account:
      type: object
      required:
        - userId
        - isPro
        - isAdmin
        - imagesGeneratedThisMonth
        - monthlyImageLimit
        - topicCredits
        - freeBreakdownsUsedThisMonth
        - freeBreakdownsPerMonth
      properties:
        userId:
          type: string
        email:
          type: ["string", "null"]
        isPro:
          type: boolean
        isAdmin:
          type: boolean
        imagesGeneratedThisMonth:
          type: integer
        monthlyImageLimit:
          type: integer
        topicCredits:
          type: integer
        freeBreakdownsUsedThisMonth:
          type: integer
        freeBreakdownsPerMonth:
          type: integer
        subscriptionStatus:
          type: ["string", "null"]
        subscriptionCurrentPeriodEnd:
          type: ["string", "null"]
          format: date-time
        aiStatus:
          type: string
          enum: [healthy, degraded, open]
          description: Tiny summary of xAI service health from the circuit breaker (for Pro users who already fetch /me)
    ImageInput:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          minLength: 1
    GeneratedImage:
      type: object
      required: [url, imagesGeneratedThisMonth]
      properties:
        url:
          type: string
        imagesGeneratedThisMonth:
          type: integer
    StripeRedirect:
      type: object
      required: [url]
      properties:
        url:
          type: string
    BreakdownInput:
      type: object
      required: [topic]
      properties:
        topic:
          type: string
          minLength: 1
    GapsInput:
      type: object
      required: [topic, breakdownTitles]
      properties:
        topic:
          type: string
          minLength: 1
        breakdownTitles:
          type: array
          items:
            type: string
    FindMoreCompaniesInput:
      type: object
      required:
        - topic
        - gapTitle
        - gapWhyExists
        - gapInnovationPotential
        - existingTickers
      properties:
        topic:
          type: string
          minLength: 1
        gapTitle:
          type: string
          minLength: 1
        gapWhyExists:
          type: string
        gapInnovationPotential:
          type: string
        existingTickers:
          type: array
          items:
            type: string
    FindMoreCompaniesResult:
      type: object
      required: [companies]
      properties:
        companies:
          type: array
          items:
            type: object
            required: [name, ticker, exchange, relevance]
            properties:
              name:
                type: string
              ticker:
                type: string
              exchange:
                type: string
              relevance:
                type: string
    StockAnalysisInput:
      type: object
      required:
        - name
        - ticker
        - exchange
        - relevance
        - topic
        - gapTitle
        - gapWhyExists
        - gapInnovationPotential
      properties:
        name:
          type: string
          minLength: 1
        ticker:
          type: string
          minLength: 1
        exchange:
          type: string
          minLength: 1
        relevance:
          type: string
        topic:
          type: string
          minLength: 1
        gapTitle:
          type: string
          minLength: 1
        gapWhyExists:
          type: string
        gapInnovationPotential:
          type: string
    StockAnalysis:
      type: object
      required: [analysis]
      properties:
        analysis:
          type: string
    StripeCheckoutRequest:
      type: object
      properties:
        interval:
          type: string
          enum: [month, year]
          default: month
    CreditCheckoutRequest:
      type: object
      required: [pack]
      properties:
        pack:
          type: string
          enum: ["1", "5", "10", "20"]